refactor(core): move Store from features/store/ to src/store/
