+++ packages/core/index.ts
@@ -23,7 +23,7 @@ export type {Signal, Computed, Event, UseHookFactory} from './src/shared/signals
 export {setUseHookFactory, getUseHookFactory, effect, event, signal, computed, watch, batch} from './src/shared/signals'
 
 // Store
-export {Store} from './src/features/store'
+export {Store} from './src/store'
 export type {Slot, MarkSlot, OverlaySlot} from './src/features/slots'
 
 // Overlay
