+++ packages/core/src/features/arrownav/ArrowNavFeature.ts
@@ -1,7 +1,7 @@
 import {KEYBOARD} from '../../shared/constants'
+import type {Store} from '../../store/Store'
 import {shiftFocusPrev, shiftFocusNext} from '../navigation'
 import {selectAllText} from '../selection'
-import type {Store} from '../store/Store'
 
 export class ArrowNavFeature {
 	#keydownHandler?: (e: KeyboardEvent) => void
