+++ packages/core/src/features/editing/utils/deleteMark.ts
@@ -1,6 +1,6 @@
 import {childAt} from '../../../shared/checkers'
+import type {Store} from '../../../store'
 import {toString} from '../../parsing'
-import type {Store} from '../../store'
 
 export function deleteMark(place: 'prev' | 'self' | 'next', store: Store) {
 	const placeMap = {
